requirements.txt
robotframework
robotframework-seleniumlibrary
robotframework-requests
deta
selenium==4.9